API-CRUD/app/schemas/product.py
from pydantic import BaseModel

class ProductCreate(BaseModel):
    name: str
    description: str
    price: float
    stock: int

class ProductUpdate(BaseModel):
    name: str = None
    description: str = None
    price: float = None
    stock: int = None

class ProductOut(ProductCreate):
    id: int

    class Config:
        orm_mode = True
