API-CRUD/requirements.txt
fastapi==0.95.0
sqlmodel==0.0.8
uvicorn==0.21.1
python-jose==3.3.0
alembic==1.10.1
